+++ core/ui/src/main/res/drawable/ic_signup_complete_etc.xml
