+++ core/ui/src/main/res/drawable/ic_signup_jobseeker.xml
