+++ feature/mypage/src/main/java/com/example/sachosaeng/feature/mypage/modify/ModifiyUserInfoUiState.kt
@@ -1,0 +1,9 @@
+package com.example.sachosaeng.feature.mypage.modify
+
+import com.example.sachosaeng.core.ui.UserType
+
+data class ModifiyUserInfoUiState (
+    val userName: String,
+    val userType: UserType,
+    val canSave: Boolean
+)
