+++ feature/mypage/src/main/res/drawable/ic_warning_black_small.xml
