--- feature/src/main/java/com/example/sachosaeng/feature/mypage/modify/ModifiyUserInfoUiState.kt
@@ -1,9 +1,0 @@
-package com.example.sachosaeng.feature.mypage.modify
-
-import com.example.sachosaeng.feature.signup.selectusertype.UserType
-
-data class ModifiyUserInfoUiState (
-    val userName: String,
-    val userType: UserType,
-    val canSave: Boolean
-)
