+++ feature/src/main/java/com/example/sachosaeng/feature/signup/selectusertype/SelectUserTypeViewModel.kt
@@ -1,7 +1,7 @@
 package com.example.sachosaeng.feature.signup.selectusertype
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
+import com.example.sachosaeng.core.ui.UserType
 import dagger.hilt.android.lifecycle.HiltViewModel
 import org.orbitmvi.orbit.Container
 import org.orbitmvi.orbit.ContainerHost
