+++ feature/src/main/java/com/example/sachosaeng/feature/signup/signupcomplete/SignUpCompleteUiState.kt
@@ -1,6 +1,6 @@
 package com.example.sachosaeng.feature.signup.signupcomplete
 
-import com.example.sachosaeng.feature.signup.selectusertype.UserType
+import com.example.sachosaeng.core.ui.UserType
 
 data class SignUpCompleteUiState (
     val userType : UserType = UserType.NEWCOMER,
