+++ settings.gradle.kts
@@ -24,3 +24,5 @@ include(":core:ui")
 include(":core:util")
 include(":feature:auth")
 include(":feature:home")
+include(":feature:mypage")
+include(":feature:webview")
